renamed postCard to the more fitting name PostDataView in order to differentiate it from the actual postCard
