+++ frontend/src/components/PostDataView.vue
